+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Fix auto-increment of open loan count in patron view. Fixes UICHKOUT-28.
 * Use more-current stripes-components. Refs STRIPES-495.
 * Display loan policy associated with loan on check out. Fixes UICHKOUT-37.
+* Get rolling loan period from loan policy. Fixes UICHKOUT-25.
 
 ## [1.1.2](https://github.com/folio-org/ui-checkout/tree/v1.1.2) (2017-09-02)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v1.1.1...v1.1.2)